ref: cloning integration tests folder structure
