+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/export-config.scenario-piece-exporter.e2e-spec.ts
@@ -19,7 +19,7 @@ import {
   DeleteProjectAndOrganization,
   GivenScenarioExists,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
