+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/marxan-execution-metadata.piece-exporter.e2e-spec.ts
@@ -17,7 +17,7 @@ import {
   DeleteMarxanExecutionMetadata,
   GivenMarxanExecutionMetadata,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
