+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/planning-area-gadm.piece-exporter.e2e-spec.ts
@@ -1,33 +1,26 @@
+import { PlanningAreaGadmPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/planning-area-gadm.piece-exporter';
 import { geoprocessingConnections } from '@marxan-geoprocessing/ormconfig';
 import { ClonePiece, ExportJobInput } from '@marxan/cloning';
 import { ResourceKind } from '@marxan/cloning/domain';
+import { PlanningAreaGadmContent } from '@marxan/cloning/infrastructure/clone-piece-data/planning-area-gadm';
 import { FileRepository, FileRepositoryModule } from '@marxan/files-repository';
-import { PlanningUnitGridShape } from '@marxan/scenarios-planning-unit';
 import { FixtureType } from '@marxan/utils/tests/fixture-type';
 import { Logger } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { getEntityManagerToken, TypeOrmModule } from '@nestjs/typeorm';
 import { isLeft, Right } from 'fp-ts/lib/Either';
 import { Readable } from 'stream';
-import { EntityManager, In, Repository } from 'typeorm';
+import { EntityManager } from 'typeorm';
 import { v4 } from 'uuid';
 import {
-  PlanningUnitsGeom,
-  ProjectsPuEntity,
-} from '@marxan-jobs/planning-unit-geometry';
-import { Polygon } from 'geojson';
-import { PlanningUnitsGridGeojsonPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/planning-units-grid-geojson.piece-exporter';
-import {
   DeleteProjectAndOrganization,
-  DeleteProjectPus,
   GivenProjectExists,
-  GivenProjectPus,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
 
-describe(PlanningUnitsGridGeojsonPieceExporter, () => {
+describe(PlanningAreaGadmPieceExporter, () => {
   beforeEach(async () => {
     fixtures = await getFixtures();
   }, 10_000);
@@ -36,13 +29,18 @@ describe(PlanningUnitsGridGeojsonPieceExporter, () => {
     await fixtures?.cleanUp();
   });
 
-  it('should save succesfully project planning units grid geojson', async () => {
-    const input = fixtures.GivenAPlanningUnitsGridExportJob();
-    await fixtures.GivenProjectExist();
-    await fixtures.GivenPlanningUnits();
+  it('should throw when project is not found', async () => {
+    const input = fixtures.GivenAPlanningAreaGadmExportJob();
     await fixtures
       .WhenPieceExporterIsInvoked(input)
-      .ThenAPlanningUnitsGridFileIsSaved();
+      .ThenAProjectExistErrorShouldBeThrown();
+  });
+  it('should save file succesfully when project is found', async () => {
+    const input = fixtures.GivenAPlanningAreaGadmExportJob();
+    await fixtures.GivenProjectWithGadmArea();
+    await fixtures
+      .WhenPieceExporterIsInvoked(input)
+      .ThenPlanningAreaGadmFileIsSaved();
   });
 });
 
@@ -54,88 +52,79 @@ const getFixtures = async () => {
         keepConnectionAlive: true,
         logging: false,
       }),
-      TypeOrmModule.forRoot({
-        ...geoprocessingConnections.default,
-        keepConnectionAlive: true,
-        logging: false,
-      }),
-      TypeOrmModule.forFeature([ProjectsPuEntity, PlanningUnitsGeom]),
       FileRepositoryModule,
     ],
     providers: [
-      PlanningUnitsGridGeojsonPieceExporter,
+      PlanningAreaGadmPieceExporter,
       { provide: Logger, useValue: { error: () => {}, setContext: () => {} } },
     ],
   }).compile();
 
   await sandbox.init();
   const projectId = v4();
-  const otherProjectId = v4();
   const organizationId = v4();
-  const sut = sandbox.get(PlanningUnitsGridGeojsonPieceExporter);
+  const sut = sandbox.get(PlanningAreaGadmPieceExporter);
   const apiEntityManager: EntityManager = sandbox.get(
     getEntityManagerToken(geoprocessingConnections.apiDB),
   );
-  const geoEntityManager: EntityManager = sandbox.get(
-    getEntityManagerToken(geoprocessingConnections.default),
-  );
   const fileRepository = sandbox.get(FileRepository);
+  const expectedContent: PlanningAreaGadmContent = {
+    l1: 'NAM.12_1',
+    l2: 'NAM.12.7_1',
+    planningUnitAreakm2: 500,
+    bbox: [10, 11, 12, 13],
+    country: 'AGO',
+  };
 
   return {
     cleanUp: async () => {
-      await DeleteProjectAndOrganization(
+      return DeleteProjectAndOrganization(
         apiEntityManager,
         projectId,
         organizationId,
       );
-      await DeleteProjectPus(geoEntityManager, projectId);
-      return DeleteProjectPus(geoEntityManager, otherProjectId);
     },
-    GivenAPlanningUnitsGridExportJob: (): ExportJobInput => {
+    GivenAPlanningAreaGadmExportJob: (): ExportJobInput => {
       return {
         allPieces: [
           { resourceId: projectId, piece: ClonePiece.ProjectMetadata },
-          {
-            resourceId: projectId,
-            piece: ClonePiece.PlanningUnitsGridGeojson,
-          },
+          { resourceId: projectId, piece: ClonePiece.PlanningAreaGAdm },
         ],
         componentId: v4(),
         exportId: v4(),
-        piece: ClonePiece.PlanningUnitsGridGeojson,
+        piece: ClonePiece.PlanningAreaGAdm,
         resourceId: projectId,
         resourceKind: ResourceKind.Project,
         isCloning: false,
       };
     },
-    GivenProjectExist: async () => {
+    GivenProjectWithGadmArea: async () => {
       return GivenProjectExists(apiEntityManager, projectId, organizationId, {
-        planning_unit_grid_shape: PlanningUnitGridShape.FromShapefile,
-        planning_unit_area_km2: 30,
+        description: 'desc',
+        admin_area_l1_id: 'NAM.12_1',
+        admin_area_l2_id: 'NAM.12.7_1',
+        planning_unit_area_km2: 500,
         bbox: JSON.stringify([10, 11, 12, 13]),
+        country_id: 'AGO',
       });
-    },
-    GivenPlanningUnits: async () => {
-      await GivenProjectPus(geoEntityManager, projectId, 2);
-
-      return GivenProjectPus(geoEntityManager, otherProjectId, 1);
     },
     WhenPieceExporterIsInvoked: (input: ExportJobInput) => {
       return {
-        ThenAPlanningUnitsGridFileIsSaved: async () => {
+        ThenAProjectExistErrorShouldBeThrown: async () => {
+          await expect(sut.run(input)).rejects.toThrow(
+            /Gadm data not found for project with ID/gi,
+          );
+        },
+        ThenPlanningAreaGadmFileIsSaved: async () => {
           const result = await sut.run(input);
           const file = await fileRepository.get(result.uris[0].uri);
           expect((file as Right<Readable>).right).toBeDefined();
           if (isLeft(file)) throw new Error();
           const savedStrem = file.right;
-          const content = await readSavedFile<{
-            type: string;
-            bbox: number[];
-            coordinates: Polygon[];
-          }>(savedStrem);
-          expect(content.type).toBe('MultiPolygon');
-          expect(content.bbox).toEqual([10, 11, 12, 13]);
-          expect(content.coordinates).toHaveLength(2);
+          const content = await readSavedFile<PlanningAreaGadmContent>(
+            savedStrem,
+          );
+          expect(content).toEqual(expectedContent);
         },
       };
     },
