+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/planning-units-grid.piece-exporter.e2e-spec.ts
@@ -22,7 +22,7 @@ import {
   GivenProjectExists,
   GivenProjectPus,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
