+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/project-custom-features.piece-exporter.e2e-spec.ts
@@ -1,29 +1,29 @@
+import { ProjectCustomFeaturesPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/project-custom-features.piece-exporter';
 import { geoprocessingConnections } from '@marxan-geoprocessing/ormconfig';
 import { ClonePiece, ExportJobInput } from '@marxan/cloning';
 import { ResourceKind } from '@marxan/cloning/domain';
+import { ProjectCustomFeaturesContent } from '@marxan/cloning/infrastructure/clone-piece-data/project-custom-features';
 import { FileRepository, FileRepositoryModule } from '@marxan/files-repository';
+import { GeoFeatureGeometry } from '@marxan/geofeatures';
 import { FixtureType } from '@marxan/utils/tests/fixture-type';
 import { Logger } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { getEntityManagerToken, TypeOrmModule } from '@nestjs/typeorm';
 import { isLeft, Right } from 'fp-ts/lib/Either';
 import { Readable } from 'stream';
-import { EntityManager, In, Repository } from 'typeorm';
+import { EntityManager, In } from 'typeorm';
 import { v4 } from 'uuid';
-import { ProjectCustomProtectedAreasPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/project-custom-protected-areas.piece-exporter';
-import { ProtectedArea } from '@marxan/protected-areas';
-import { ProjectCustomProtectedAreasContent } from '@marxan/cloning/infrastructure/clone-piece-data/project-custom-protected-areas';
 import {
   DeleteProjectAndOrganization,
-  DeleteProtectedAreas,
-  GivenCustomProtectedAreas,
+  GivenFeatures,
+  GivenFeaturesData,
   GivenProjectExists,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
 
-describe(ProjectCustomProtectedAreasPieceExporter, () => {
+describe(ProjectCustomFeaturesPieceExporter, () => {
   beforeEach(async () => {
     fixtures = await getFixtures();
   }, 10_000);
@@ -32,21 +32,22 @@ describe(ProjectCustomProtectedAreasPieceExporter, () => {
     await fixtures?.cleanUp();
   });
 
-  it('should save empty file when there are not any project custom protected areas', async () => {
-    const input = fixtures.GivenAProjectCustomProtectedAreasExportJob();
+  it("saves an empty file when project doesn't have custom features", async () => {
+    const input = fixtures.GivenAProjectCustomFeaturesExportJob();
     await fixtures.GivenProjectExist();
-    await fixtures.GivenNoCustomProtectedAreaForProject();
+    fixtures.GivenNoCustomFeaturesForProject();
     await fixtures
       .WhenPieceExporterIsInvoked(input)
-      .ThenAnEmptyProjectCustomProtectedAreasFileIsSaved();
+      .ThenAnEmptyProjectCustomFeaturesFileIsSaved();
   });
-  it('should save file succesfully when there are project custom protected areas', async () => {
-    const input = fixtures.GivenAProjectCustomProtectedAreasExportJob();
+
+  it('saves succesfully features data when the project has custom features', async () => {
+    const input = fixtures.GivenAProjectCustomFeaturesExportJob();
     await fixtures.GivenProjectExist();
-    await fixtures.GivenCustomProtectedAreaForProject();
+    await fixtures.GivenCustomFeaturesForProject();
     await fixtures
       .WhenPieceExporterIsInvoked(input)
-      .ThenAProjectCustomProtectedAreasFileIsSaved();
+      .ThenAProjectCustomFeaturesFileIsSaved();
   });
 });
 
@@ -63,29 +64,31 @@ const getFixtures = async () => {
         keepConnectionAlive: true,
         logging: false,
       }),
-      TypeOrmModule.forFeature([ProtectedArea]),
+      TypeOrmModule.forFeature([GeoFeatureGeometry]),
       FileRepositoryModule,
     ],
     providers: [
-      ProjectCustomProtectedAreasPieceExporter,
+      ProjectCustomFeaturesPieceExporter,
       { provide: Logger, useValue: { error: () => {}, setContext: () => {} } },
     ],
   }).compile();
 
   await sandbox.init();
   const projectId = v4();
-  const otherProjectId = v4();
-  let projectPlanningAreaId: string;
-  let otherProjectPlanningAreaId: string;
   const organizationId = v4();
-  const sut = sandbox.get(ProjectCustomProtectedAreasPieceExporter);
+  const sut = sandbox.get(ProjectCustomFeaturesPieceExporter);
   const apiEntityManager: EntityManager = sandbox.get(
     getEntityManagerToken(geoprocessingConnections.apiDB),
   );
   const geoEntityManager: EntityManager = sandbox.get(
     getEntityManagerToken(geoprocessingConnections.default),
   );
+  const featuresDataRepo = geoEntityManager.getRepository(GeoFeatureGeometry);
   const fileRepository = sandbox.get(FileRepository);
+
+  let featureIds: string[] = [];
+  const amountOfCustomFeatures = 5;
+  const recordsOfDataForEachCustomFeature = 3;
 
   return {
     cleanUp: async () => {
@@ -94,70 +97,69 @@ const getFixtures = async () => {
         projectId,
         organizationId,
       );
-      return DeleteProtectedAreas(geoEntityManager, [
-        projectPlanningAreaId,
-        otherProjectPlanningAreaId,
-      ]);
+      await featuresDataRepo.delete({ featureId: In(featureIds) });
     },
-    GivenAProjectCustomProtectedAreasExportJob: (): ExportJobInput => {
+    GivenAProjectCustomFeaturesExportJob: (): ExportJobInput => {
       return {
         allPieces: [
           { resourceId: projectId, piece: ClonePiece.ProjectMetadata },
           {
             resourceId: projectId,
-            piece: ClonePiece.ProjectCustomProtectedAreas,
+            piece: ClonePiece.ProjectCustomFeatures,
           },
         ],
         componentId: v4(),
         exportId: v4(),
-        piece: ClonePiece.ProjectMetadata,
+        piece: ClonePiece.ProjectCustomFeatures,
         resourceId: projectId,
         resourceKind: ResourceKind.Project,
         isCloning: false,
       };
     },
     GivenProjectExist: async () => {
-      return GivenProjectExists(apiEntityManager, projectId, organizationId, {
-        description: 'desc',
-        planning_unit_area_km2: 500,
-      });
+      return GivenProjectExists(apiEntityManager, projectId, organizationId);
     },
-    GivenNoCustomProtectedAreaForProject: async () => {
-      otherProjectPlanningAreaId = (
-        await GivenCustomProtectedAreas(geoEntityManager, 1, otherProjectId)
-      )[0].id;
-    },
-    GivenCustomProtectedAreaForProject: async () => {
-      projectPlanningAreaId = (
-        await GivenCustomProtectedAreas(geoEntityManager, 1, projectId)
-      )[0].id;
+    GivenNoCustomFeaturesForProject: () => {},
+    GivenCustomFeaturesForProject: async () => {
+      const { customFeatures } = await GivenFeatures(
+        apiEntityManager,
+        0,
+        amountOfCustomFeatures,
+        projectId,
+      );
+      featureIds = customFeatures.map((feature) => feature.id);
+      await GivenFeaturesData(
+        geoEntityManager,
+        recordsOfDataForEachCustomFeature,
+        featureIds,
+      );
     },
     WhenPieceExporterIsInvoked: (input: ExportJobInput) => {
       return {
-        ThenAnEmptyProjectCustomProtectedAreasFileIsSaved: async () => {
+        ThenAnEmptyProjectCustomFeaturesFileIsSaved: async () => {
           const result = await sut.run(input);
           const file = await fileRepository.get(result.uris[0].uri);
           expect((file as Right<Readable>).right).toBeDefined();
           if (isLeft(file)) throw new Error();
           const savedStrem = file.right;
-          const content = await readSavedFile(savedStrem);
-          expect(content).toEqual([]);
-        },
-        ThenAProjectCustomProtectedAreasFileIsSaved: async () => {
-          const result = await sut.run(input);
-          const file = await fileRepository.get(result.uris[0].uri);
-          expect((file as Right<Readable>).right).toBeDefined();
-          if (isLeft(file)) throw new Error();
-          const savedStrem = file.right;
-          const content = (
-            await readSavedFile<ProjectCustomProtectedAreasContent[]>(
-              savedStrem,
-            )
-          )[0];
-          expect(content.fullName).toBe(
-            `custom protected area 1 of ${projectId}`,
+          const content = await readSavedFile<ProjectCustomFeaturesContent>(
+            savedStrem,
           );
-          expect(content.ewkb.length).toBeGreaterThan(0);
+          expect(content.features).toEqual([]);
+        },
+        ThenAProjectCustomFeaturesFileIsSaved: async () => {
+          const result = await sut.run(input);
+          const file = await fileRepository.get(result.uris[0].uri);
+          expect((file as Right<Readable>).right).toBeDefined();
+          if (isLeft(file)) throw new Error();
+          const savedStrem = file.right;
+          const content = await readSavedFile<ProjectCustomFeaturesContent>(
+            savedStrem,
+          );
+          expect(content.features).toHaveLength(amountOfCustomFeatures);
+          const { feature_class_name, data } = content.features[0];
+          expect(feature_class_name).toEqual(`custom-${projectId}-1`);
+          expect(data).toHaveLength(recordsOfDataForEachCustomFeature);
         },
       };
     },
