+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/project-metadata.piece-exporter.e2e-spec.ts
@@ -1,7 +1,9 @@
+import { ProjectMetadataPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/project-metadata.piece-exporter';
 import { geoprocessingConnections } from '@marxan-geoprocessing/ormconfig';
 import { ClonePiece, ExportJobInput } from '@marxan/cloning';
 import { ResourceKind } from '@marxan/cloning/domain';
 import { FileRepository, FileRepositoryModule } from '@marxan/files-repository';
+import { PlanningUnitGridShape } from '@marxan/scenarios-planning-unit';
 import { FixtureType } from '@marxan/utils/tests/fixture-type';
 import { Logger } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
@@ -10,17 +12,16 @@ import { isLeft, Right } from 'fp-ts/lib/Either';
 import { Readable } from 'stream';
 import { EntityManager } from 'typeorm';
 import { v4 } from 'uuid';
-import { ScenarioMetadataPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/scenario-metadata.piece-exporter';
-import { ScenarioMetadataContent } from '@marxan/cloning/infrastructure/clone-piece-data/scenario-metadata';
+import { ProjectMetadataContent } from '@marxan/cloning/infrastructure/clone-piece-data/project-metadata';
 import {
   DeleteProjectAndOrganization,
-  GivenScenarioExists,
+  GivenProjectExists,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
 
-describe(ScenarioMetadataPieceExporter, () => {
+describe(ProjectMetadataPieceExporter, () => {
   beforeEach(async () => {
     fixtures = await getFixtures();
   }, 10_000);
@@ -29,18 +30,18 @@ describe(ScenarioMetadataPieceExporter, () => {
     await fixtures?.cleanUp();
   });
 
-  it('should throw when the scenario is not found', async () => {
-    const input = fixtures.GivenAScenarioMetadataExportJob();
+  it('should throw when project is not found', async () => {
+    const input = fixtures.GivenAProjectMetadataExportJob();
     await fixtures
       .WhenPieceExporterIsInvoked(input)
-      .ThenASceanarioExistErrorShouldBeThrown();
+      .ThenAProjectExistErrorShouldBeThrown();
   });
-  it('should save file succesfully when the scenario is found', async () => {
-    const input = fixtures.GivenAScenarioMetadataExportJob();
-    await fixtures.GivenScenarioExist();
+  it('should save file succesfully when project is found', async () => {
+    const input = fixtures.GivenAProjectMetadataExportJob();
+    await fixtures.GivenProjectExist();
     await fixtures
       .WhenPieceExporterIsInvoked(input)
-      .ThenScenarioMetadataFileIsSaved();
+      .ThenProjectMetadataFileIsSaved();
   });
 });
 
@@ -55,26 +56,23 @@ const getFixtures = async () => {
       FileRepositoryModule,
     ],
     providers: [
-      ScenarioMetadataPieceExporter,
+      ProjectMetadataPieceExporter,
       { provide: Logger, useValue: { error: () => {}, setContext: () => {} } },
     ],
   }).compile();
 
   await sandbox.init();
   const projectId = v4();
-  const scenarioId = v4();
   const organizationId = v4();
-  const sut = sandbox.get(ScenarioMetadataPieceExporter);
+  const sut = sandbox.get(ProjectMetadataPieceExporter);
   const apiEntityManager: EntityManager = sandbox.get(
     getEntityManagerToken(geoprocessingConnections.apiDB),
   );
   const fileRepository = sandbox.get(FileRepository);
-  const expectedContent: ScenarioMetadataContent = {
-    name: `test scenario - ${scenarioId}`,
-    description: 'desc',
-    blm: 1,
-    numberOfRuns: 6,
-    metadata: { marxanInputParameterFile: { meta: '1' } },
+  const expectedContent: ProjectMetadataContent = {
+    name: `test project - ${projectId}`,
+    planningUnitGridShape: PlanningUnitGridShape.Square,
+    projectAlreadyCreated: false,
   };
 
   return {
@@ -85,48 +83,37 @@ const getFixtures = async () => {
         organizationId,
       );
     },
-    GivenAScenarioMetadataExportJob: (): ExportJobInput => {
+    GivenAProjectMetadataExportJob: (): ExportJobInput => {
       return {
         allPieces: [
-          { resourceId: scenarioId, piece: ClonePiece.ScenarioMetadata },
+          { resourceId: projectId, piece: ClonePiece.ProjectMetadata },
         ],
         componentId: v4(),
         exportId: v4(),
-        piece: ClonePiece.ScenarioMetadata,
-        resourceId: scenarioId,
-        resourceKind: ResourceKind.Scenario,
+        piece: ClonePiece.ProjectMetadata,
+        resourceId: projectId,
+        resourceKind: ResourceKind.Project,
         isCloning: false,
       };
     },
-    GivenScenarioExist: async () => {
-      return GivenScenarioExists(
-        apiEntityManager,
-        scenarioId,
-        projectId,
-        organizationId,
-        {
-          description: 'desc',
-          blm: 1,
-          number_of_runs: 6,
-          metadata: { marxanInputParameterFile: { meta: '1' } },
-        },
-      );
+    GivenProjectExist: async () => {
+      return GivenProjectExists(apiEntityManager, projectId, organizationId);
     },
     WhenPieceExporterIsInvoked: (input: ExportJobInput) => {
       return {
-        ThenASceanarioExistErrorShouldBeThrown: async () => {
+        ThenAProjectExistErrorShouldBeThrown: async () => {
           await expect(sut.run(input)).rejects.toThrow(/does not exist/gi);
         },
-        ThenScenarioMetadataFileIsSaved: async () => {
+        ThenProjectMetadataFileIsSaved: async () => {
           const result = await sut.run(input);
           const file = await fileRepository.get(result.uris[0].uri);
           expect((file as Right<Readable>).right).toBeDefined();
           if (isLeft(file)) throw new Error();
           const savedStrem = file.right;
-          const content = await readSavedFile<ScenarioMetadataContent>(
+          const content = await readSavedFile<ProjectMetadataContent>(
             savedStrem,
           );
-          expect(content).toEqual(expectedContent);
+          expect(content).toMatchObject(expectedContent);
         },
       };
     },
