+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/scenario-features-data.piece-exporter.e2e-spec.ts
@@ -20,7 +20,7 @@ import {
   GivenOutputScenarioFeaturesData,
   GivenScenarioExists,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
