+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/scenario-planning-units-data.piece-exporter.e2e-spec.ts
@@ -4,6 +4,7 @@ import { ResourceKind } from '@marxan/cloning/domain';
 import { FileRepository, FileRepositoryModule } from '@marxan/files-repository';
 import {
   LockStatus,
+  ScenariosPuCostDataGeo,
   ScenariosPuPaDataGeo,
 } from '@marxan/scenarios-planning-unit';
 import { FixtureType } from '@marxan/utils/tests/fixture-type';
@@ -12,31 +13,26 @@ import { Test } from '@nestjs/testing';
 import { getEntityManagerToken, TypeOrmModule } from '@nestjs/typeorm';
 import { isLeft, Right } from 'fp-ts/lib/Either';
 import { Readable } from 'stream';
-import { EntityManager, In, Repository } from 'typeorm';
+import { EntityManager } from 'typeorm';
 import { v4 } from 'uuid';
 import {
   PlanningUnitsGeom,
   ProjectsPuEntity,
 } from '@marxan-jobs/planning-unit-geometry';
-import { ScenarioRunResultsPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/scenario-run-results.piece-exporter';
-import { ScenarioRunResultsContent } from '@marxan/cloning/infrastructure/clone-piece-data/scenario-run-results';
-import { BlmFinalResultEntity } from '@marxan/blm-calibration';
-import { OutputScenariosPuDataGeoEntity } from '@marxan/marxan-output';
+import { ScenarioPlanningUnitsDataPieceExporter } from '@marxan-geoprocessing/export/pieces-exporters/scenario-planning-units-data.piece-exporter';
+import { ScenarioPlanningUnitsDataContent } from '@marxan/cloning/infrastructure/clone-piece-data/scenario-planning-units-data';
 import {
   DeleteProjectAndOrganization,
   DeleteProjectPus,
-  DeleteScenarioBlmResults,
-  DeleteScenarioOutputPuDataResults,
-  GivenScenarioBlmResults,
   GivenScenarioExists,
-  GivenScenarioOutputPuData,
+  GivenScenarioPuCostData,
   GivenScenarioPuData,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
 
-describe(ScenarioRunResultsPieceExporter, () => {
+describe(ScenarioPlanningUnitsDataPieceExporter, () => {
   beforeEach(async () => {
     fixtures = await getFixtures();
   }, 10_000);
@@ -45,24 +41,13 @@ describe(ScenarioRunResultsPieceExporter, () => {
     await fixtures?.cleanUp();
   });
 
-  it('should save empty scenario run results file', async () => {
-    const input = fixtures.GivenAScenarioRunResultsExportJob();
+  it('should save succesfully scenario planning units data', async () => {
+    const input = fixtures.GivenAScenarioPlanningUnistDataExportJob();
     await fixtures.GivenScenarioExist();
-    await fixtures.GivenScenarioPlanningUnits();
-    await fixtures.GivenNoScenarioRunResulst();
+    await fixtures.GivenPlanningUnitsData();
     await fixtures
       .WhenPieceExporterIsInvoked(input)
-      .ThenAnEmptyScenarioRunResultsFileIsSaved();
-  });
-
-  it('should save succesfully scenario run results', async () => {
-    const input = fixtures.GivenAScenarioRunResultsExportJob();
-    await fixtures.GivenScenarioExist();
-    await fixtures.GivenScenarioPlanningUnits();
-    await fixtures.GivenScenarioRunResulst();
-    await fixtures
-      .WhenPieceExporterIsInvoked(input)
-      .ThenAScenarioRunResultsFileIsSaved();
+      .ThenAScenarioPlanningUnitsDataFileIsSaved();
   });
 });
 
@@ -83,13 +68,12 @@ const getFixtures = async () => {
         ProjectsPuEntity,
         PlanningUnitsGeom,
         ScenariosPuPaDataGeo,
-        BlmFinalResultEntity,
-        OutputScenariosPuDataGeoEntity,
+        ScenariosPuCostDataGeo,
       ]),
       FileRepositoryModule,
     ],
     providers: [
-      ScenarioRunResultsPieceExporter,
+      ScenarioPlanningUnitsDataPieceExporter,
       { provide: Logger, useValue: { error: () => {}, setContext: () => {} } },
     ],
   }).compile();
@@ -98,8 +82,7 @@ const getFixtures = async () => {
   const projectId = v4();
   const scenarioId = v4();
   const organizationId = v4();
-  let scenarioPus: ScenariosPuPaDataGeo[];
-  const sut = sandbox.get(ScenarioRunResultsPieceExporter);
+  const sut = sandbox.get(ScenarioPlanningUnitsDataPieceExporter);
   const apiEntityManager: EntityManager = sandbox.get(
     getEntityManagerToken(geoprocessingConnections.apiDB),
   );
@@ -113,25 +96,22 @@ const getFixtures = async () => {
       await DeleteProjectAndOrganization(
         apiEntityManager,
         projectId,
-        scenarioId,
+        organizationId,
       );
-      await DeleteScenarioOutputPuDataResults(geoEntityManager, scenarioId);
-
-      await DeleteProjectPus(geoEntityManager, projectId);
-      return DeleteScenarioBlmResults(geoEntityManager, scenarioId);
+      return DeleteProjectPus(geoEntityManager, projectId);
     },
-    GivenAScenarioRunResultsExportJob: (): ExportJobInput => {
+    GivenAScenarioPlanningUnistDataExportJob: (): ExportJobInput => {
       return {
         allPieces: [
           { resourceId: projectId, piece: ClonePiece.ProjectMetadata },
           {
             resourceId: scenarioId,
-            piece: ClonePiece.ScenarioRunResults,
+            piece: ClonePiece.ScenarioPlanningUnitsData,
           },
         ],
         componentId: v4(),
         exportId: v4(),
-        piece: ClonePiece.ScenarioRunResults,
+        piece: ClonePiece.ScenarioPlanningUnitsData,
         resourceId: scenarioId,
         resourceKind: ResourceKind.Project,
         isCloning: false,
@@ -145,8 +125,8 @@ const getFixtures = async () => {
         organizationId,
       );
     },
-    GivenScenarioPlanningUnits: async () => {
-      scenarioPus = await GivenScenarioPuData(
+    GivenPlanningUnitsData: async () => {
+      const scenariPus = await GivenScenarioPuData(
         geoEntityManager,
         scenarioId,
         projectId,
@@ -159,45 +139,33 @@ const getFixtures = async () => {
           yloc: 8,
         },
       );
+
+      return GivenScenarioPuCostData(geoEntityManager, scenariPus);
     },
-    GivenScenarioRunResulst: async () => {
-      await GivenScenarioBlmResults(geoEntityManager, scenarioId);
-      return GivenScenarioOutputPuData(geoEntityManager, scenarioPus);
-    },
-    GivenNoScenarioRunResulst: async (): Promise<void> => {},
     WhenPieceExporterIsInvoked: (input: ExportJobInput) => {
       return {
-        ThenAnEmptyScenarioRunResultsFileIsSaved: async () => {
+        ThenAScenarioPlanningUnitsDataFileIsSaved: async () => {
           const result = await sut.run(input);
           const file = await fileRepository.get(result.uris[0].uri);
           expect((file as Right<Readable>).right).toBeDefined();
           if (isLeft(file)) throw new Error();
           const savedStrem = file.right;
-          const content = await readSavedFile(savedStrem);
-          expect(content).toEqual({ blmResults: [], marxanRunResults: [] });
-        },
-        ThenAScenarioRunResultsFileIsSaved: async () => {
-          const result = await sut.run(input);
-          const file = await fileRepository.get(result.uris[0].uri);
-          expect((file as Right<Readable>).right).toBeDefined();
-          if (isLeft(file)) throw new Error();
-          const savedStrem = file.right;
-          const content = await readSavedFile<ScenarioRunResultsContent>(
+          const content = await readSavedFile<ScenarioPlanningUnitsDataContent>(
             savedStrem,
           );
-          const expectedMarxanResults = [1, 2, 3].map((value) => ({
-            includedCount: 7,
-            values: [false, true, false, true],
-            puid: value,
-          }));
-          expect(content.blmResults[0]).toEqual({
-            blmValue: 30,
-            boundaryLength: 500,
-            cost: 1,
-          });
+          expect(content.planningUnitsData).toHaveLength(3);
           expect(
-            content.marxanRunResults.sort((a, b) => a.puid - a.puid),
-          ).toEqual(expectedMarxanResults);
+            content.planningUnitsData.every(
+              (data, index) =>
+                data.cost === 1 &&
+                data.protectedArea === 700.0045 &&
+                data.protectedByDefault &&
+                data.lockinStatus === 1 &&
+                data.xloc === 1 &&
+                data.yloc === 8 &&
+                data.puid === index + 1,
+            ),
+          ).toEqual(true);
         },
       };
     },
