+++ api/apps/geoprocessing/test/integration/cloning/piece-exporters/scenario-protected-areas.piece-exporter.e2e-spec.ts
@@ -21,7 +21,7 @@ import {
   GivenScenarioExists,
   GivenWdpaProtectedAreas,
   readSavedFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
