+++ api/apps/geoprocessing/test/integration/cloning/piece-importers/project-custom-features.piece-importer.e2e-spec.ts
@@ -29,7 +29,7 @@ import {
   GenerateRandomGeometries,
   GivenProjectExists,
   PrepareZipFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
