+++ api/apps/geoprocessing/test/integration/cloning/piece-importers/scenario-run-results.piece-importer.e2e-spec.ts
@@ -30,7 +30,7 @@ import {
   DeleteProjectPus,
   GivenScenarioPuData,
   PrepareZipFile,
-} from './fixtures';
+} from '../fixtures';
 
 let fixtures: FixtureType<typeof getFixtures>;
